If you attempt to upload a single file, error.
Case:
Ellen has tried to do this a few times - `dark-cli headshot.png`,
`dark-cli index.html` - and it fails because we'd have to special-case
our path-prefix-stripping.

I talked about two options with her:
1) Disallow that, and require that if you really do want that, you put
the file in a directory and upload that.
2) Special-case.

The downside to (2) (and the reason we went with (1)), is that it risks
encouraging users to individually upload files (`dark-cli index.html &&
dark-cli favicon.ico && dark-cli app.js`), wonder why `Serve::latest`
only works for `app.js` (the last file uploaded), and now we have a
support problem. Much better to nudge them in the direction of doing the
right thing.

https://trello.com/c/p5Woe2zP/2275-error-instead-of-silently-failing-if-a-user-tries-to-use-dark-cli-with-a-single-file-instead-of-a-directory