Get auth from login.darklang.com
(And also bump version to 0.5.0)

We are moving auth to depend on auth0, so moving dark-lang to get auth
that way keeps browser+dark-cli passwords coming from the same source.

Auth0 _usually_ demands user input (the usual oauth flow), but in this
case, we're using grant_type=password to implement a flow that doesn't
require that.

This will not work for users prior to the auth0 cutover! Because if they
use this binary, they will be attempting to  use their old (non-auth0)
passwords against the new (auth0) login service.

Since asking users to maintain multiple passwords will confuse them and
increase our support burden, I am going to refrain from deploying this
until we cut over to the new auth flow.

https://trello.com/c/MXRVhUAl/2346-update-dark-cli-to-work-with-auth0